Initial Docker configuration files
Created initial Docker and docker Compose files! Created a specific
Dockerfile for the rest servcies application, and used a public posgres
image for the database. The docker-compose.yaml file is deploys /
configures the necessary containers to perform development in.

Note: The database credentials need to match the contents of the
settings.py file. These need to securely passed in the future,