Fixed detail views and added a root api view
Updated the url patterns to (1) have proper regexes (2) specify the
name of the associated primary key field in the model class (3) assigned
names to the views so they they could be referenced elsewhere.

Updated the UnitOfIssueList and ItemList views to perform lookusp
againstthe model primary key.

Added a default view for the root of the API that list the URLS that the
item-service app can respond to.