Added https bindings for the nginx reverse proxy
Updated the nginx.conf file to add a https server section, specifically
set basic security settings and additional headers for the django app to
use ('X-Forwarded-Proto'). Updated the docker-compose file to expose port 443
so that it will accept https traffic. Temporarily baked q server
certificate and a key into the proxy image so that the server can start
a secure listener, this needs to be fixed.

Updated the django settings to rely on the 'X-Forwarded-Proto' HTTP
header being set for the reverse url lookups to be set appropriately to
provide "https://" urls. This is an inherently insecure configuration
setting but is safe used behind a reverse proxy like I'm currently
doing. At least it will be safe when I remove the http listener...