Added a script to generate self signed CAs
Since this script is in my repo, it might as well have my default
settings! It randomly created a large master password for the CA,
generates the private key, generates a public key based on limited user
input, and ensures the results are sane.

It helpfully stores the generated files in a local directory to be
easily manipulated as necessary.