Added a script to generate user certs from a CA
Added the generate-user-cert.sh script to create user certificates in
p12 format that can be imported into OSX / Windows to be used for x509
auth with a webapp. The script is not exteremely user friendly, but it
gets the job done, and verifies that the results are correct!

Added a .gitignore file to exclude local directories I have archived
generated certs in so I don't check them in accidently.