Moved the docker-compose.yaml file up a directory
This currently breaks the proxy from working, but I didnt want to change
to much in a single checkin and lose the scope of the
docker-compose.yaml file move. Updated the docker-compose file to adjust
the working directory for the supply services. Updated the
start-web-container-script.sh to detect the directory it is run from,
and switch into it so that additional scripts can be run fromr elational
paths.