Updated the rest services to listen on port 80
They previously listened on port 8000 as a holdover from manually
running the manage.py script on a local server (not as root). Updated
the start web container script to set the default port, not sure if I
will override the settings or not through command line options. Internal
docker networking should take care of multiple individual containers
listing on port 80 internally. Only the nginx proxy should be listening
on 80 externally for docker-compose.