Fixed supply-service docker-compose exposed port
the port was originally set to 8000:8000, changed it to 80:80 to line up
with the change to the docker image. This allows access to the DRF
servcies over http in development, but once I tie user auth to x509
client certs nothing will work (unless the http request spoofs the
header).

Again I will stress, this change only affects development. Port 80 will
not be exposed in production!