Fixed docker-compose volume local reference
Apparently docker-compose volumes ignore the specified context. This led
to the whole src direcotry being used as a volume during development,
which was out of sync with what the actual container build looks like.
Fixing this issue also removed the need for me to override the
entrypoint and specify a different command to run on start. This issue
would have been introduced when I moved the docker-compose file up a
directory.