Added SAN support for generated certificates
Added subject alt names to the generated server certificates since
google chrome complains that the connection is not secure.

I went through multiple iterations for getting subject alt names added
to the generated x509 server certs. Obviously creating my own CA / certs
is a little messy, but it is necessary for this demo.

I initially fixed the request to add x509 v3 extensions and add subject
alt names to the quest, but that didn't fix the issue. This stack
overflow answer had the simplest way (that I could find) to add subject
alt names to a cert:

https://stackoverflow.com/questions/7580508/getting-chrome-to-accept-self-signed-localhost-certificate/43666288#43666288

I just created a inline file with x509v3 extensions to add the subject alt
names. At this point I haven't updated the script to prompt for them,
and I'm just defaulting to "www.$(hostname -)" and "$(hostname -f)". Since I'm
generating self signed certs, I don't have to tinker with the request
and have that go the whole way though the request process.