BREAKING CHANGE - Changed item service root URL
Unfortunatley I had to change teh root URL that django listens on. I was
originally listing on "http://{fqdn}/" but all of my attempts to have
nginx do URL rewriting caused the django absolute URl computations to
break. In order to not have to add rules for each django app
(item-service, admin, health-check, etc) I added a rule for Nginx to
route a sub-url '/supply' to the django app and updated the root urls.

Basically trying to rewrite '/supply' URLS and proxy them to '/item-service'
caused individual HyperlinkedModel URLs to be calculated incorrectly.
The generated URLs would look like:

https://farnsworth.dev.simpkins.cloud/item-service/unit-of-issue/UI/

  -- instead of --

https://farnsworth.dev.simpkins.cloud/supply/item-service/unit-of-issue/UI/

I'm not a fan of having the django app containers list on '/supply'
instead of '/' but it definitely seemed like the most clean fix. this
will also make it a little more clear about which microservice the proxy
may be pointing to (e.g. with another aviation maintance service
exposed).