Updated nginx for dynamic config and to serve static content
This is checkin could probably be split up, but it's my repo so meh. The
basics of this commit are that I'm adding two new features: (1) nginx
dynamic configuration on startup and (2) serving static content from
nginx.

Updated the Dockerfile:
- Static html content will get served from /usr/local/nginx/html to
align with nginx standards (and not apache -.-)
- The nginx config is no longer baked in place.
- Updated the RUN command to set the execute permissions on scripts for
consistency.

Updated the nginx.conf template:
- Replaced the baked in hostnames for the upstream supply service fqdn
and the https server name with "template variables" that get replaced in
the stry-point.sh script.
- Nginx now proxies calls to /supply to the django containers (not from
/ anymore).
- Removed the proxy_redirect settings, It's not necessary and we're not
doing URL rewriting.
- Added a / location to serve static content! Make sure you turn on mime
types, otherwise CSS won't render!
- Turned on auto indexing for development purposes.

Updated the entry-point.sh script:
- Set the REVERSE_PROXY_FQDN and SUPPLY_SERVICE_FQDN env variables for
the DEV-LOCAL deployment type.
- Added debug output for nginx config

Not: on development workstations and server where the deployment types
is set to DEV-LOCAL, nginx defaults to using the server hostname as the
container hostname! This make sure server certs and DNS will line up
correctly.