Added initial working reverse proxy fargate config
Wow. What a pain this was today. Ran into a bunch of strange dependency
issues (e.g. the fargate service was complaining that it's target group
wasn't attached to a load balancer). Ran into issues with the fargate
cluster running in the private subnets not being able to reach ECR to
pull down images. Ran into issues with the ELB having the wrong health
check and just spinning fargate containers down. Ugh. Ran into another
issue where the nginx docker image aborts when the upstream server
doesn't exist, and will need to fix that in the future.

I'm checking this configuration in before I break it. It's ugly, but it
works.

Having cloud watch groups for fargate is critical! And
it wasn't apparent that I needed them until it was too late. Fargate is
a huge pain to debug.