Added deregistrationdelay parameter
Ugh. Deregistration delay makes development interation times painfully
slow. It is the amount of time to wait before the ELB starts to send a
message to the fargate cliutser to drain instances. There seems to be a
nother timer delay here that I'm not aware of.

This allows the consumer to set the deregistration delay for the ELB
targets. Whew it's default is set to 5 minutes, which it makes it
painfully slow to update fargate containers in development. Set teh
default to 15 seconds, which even seems to long.