Added userdata init postgres script to bastion
Added a init script passed in as base 64 user data for teh init of the
bastion host in public subnet 1. This will need to get repped ot the
other bastion.

Created a .pgpass file containing the rds connection information on the
bastion host.Probably not the best thing to do? But heck it's
development and I need my demos to go smooth.

Also installed the postgresql96 client so that I can use psql.