Added the supply fargate service and a ALB sg
Added the fargate supply service. The initial creation and deployment
was easy, but goodness I got wrapped around the security groups. Cloud
formation doesn't automatically roll back if a change causes ECS
deployments to fail (e.g. when you change a security group and break a
ELB healthcheck). What a pain in the butt.

Updated the fargate proxy service to route traffic to the ALB insrtead
of localhost! The app works end to end now!

I added a new security group for the ALB, because you need one! NLBs
dont because they just route, but in hindsight its a bit strange that
they're different.

I documented some of my issues handling deployments in the README.