Added optional STACK_PARAMS variable processing
This allows you to specify override parameters on the command line. I
added this since it's annoying to have to update teh Makefile everytime
I introduce a new parameter and have to update a stack. I'll switch over
to using a parameter file eventually.

You can use the use the STACK_PARAMS variable like so:

 > make deploy-ea-vpc STACK_PARAMS='rHostname=simpkins.cloud'