Added DNS entries to specified hoest zone
Added a parameter for the hosted zone. In my case it is a zone that is
referenced from my personal AWS account through a NS record. Added
public DNS entries for the application URL and the two bastion hosts.
Added thos entires to the output section to make it easy to connect to
the app.