Converted REVERSE_PROXY_FQDN to standard format
When the cloud formation template passed in the reverse proxy host name
(because it creates the DNS record) it is in a format that doesn't quite
match the cert name. This change lower cases the FQDN and strips off a
trailing period.

When the template passes in:

> !Sub '${pDeploymentType}.${pPublicHostedZoneName}'

The ENV variable gets set to:

> PRODUCTION.tec.simpkins.cloud.

Which the entry point now converts to:

> production.tec.simpkins.cloud