Temporarily changing the X-Forwarded-Proto SSL check
I'm not sure how I didn't se this before, but when django runs behind
a ALB, the ALB resets the X-Forwarded-Proto value from HTTPS to HTTP.
That causes the SSL check to fail and django to think the request
processing isn't secure. That causes the hypelinkedmodel serilaizers to
return URLs with http:// NOT https:// which breaks the page (since I'm
not allowing http://). This is a canary fix, and if it works I'll follow
up with additonal changes sicne this breaks the DEV-LOCAL deploys.