Baseline postgresql.conf for postgres container
Created an initial baseline for the postgresql.conf file for my new
postgres container. Adding replication to RDS requires container options
to changes, it seems easiest to override the base docker image.

I created this postgresql.conf file by cat'ing the version that is
provided in the base image. It also has tasb by defualt, so I set the
tabstop to 8 and executed :retab in vim to swicth tabs to spaces.