Added default HBA rep entry (NOT SECURE!)
To simplify the demo process, I updated the init sctipts to add a HBA
replciation entry that allows and replication client to connect in as
long as they have the right password. This makes sure I don't have to
tink with the CIDR range before the demo on a snowball with a static IP
address.