Enabled x509 client auth for the proxy
Updated the nginx.conf template to use the USER cleint cert CA and I set
the ssl_verify_client setting to optional. I added a if block to ensure
that the /supply path requires the client verification to be successful.
This allows users to see the static files (main page, css, and error
pages) but still requires them to authenticate for the app.

Also updated a setting in the nginx.conf file to have nginx intercept
django errors (e.g. 404)_and display the associated error page that is
configured in the proxy. Basically Django throws a 404, and nginx
displays it's own 404 page. This gives it a more seamless experience.

Added auto indexing for /error-pages in the nginx.conf

Created a error-pages directory. Moved the existing error pages (401 and
404) into it. Added new error pagesz (403 and 405) to cover client auth
errors.

Added a link to the index.html to the error-pages directory.