Added custom item service ./manage.py commands
Created management commands for the item service. They can be invoked by
running './manage.py [command]' as the ec2-user on the dev server. This
makes it easy to manage item service application admins on the command
line (e.g. for initial admin creation).

This checkin is occuring prior to changes I made to pass the user client
cert credentials to django and tie them to the user managment. This is
specifically useful for registering the initial admin for the application
where the user cert DN (distingiushed name) is a pain to type.

As always, nmke sure you source the django virtual env first:

 [ec2-user@farnsworth src]$ source ~/python-3.6-django-env/bin/activate

Example usage:

 [ec2-user@farnsworth src]$ ./manage.py list-admins
 [ec2-user@farnsworth src]$ ./manage.py list-non-admins
 [ec2-user@farnsworth src]$ ./manage.py create-admin <user cert dn>
 [ec2-user@farnsworth src]$ ./manage.py delete-admin <user cert dn>

Since the upcoming django authentication changes automatically register
and activate users, I added 'list-non-admins' so that you could easily
get the DN for a specific user'. The 'create-admin' script creates a
brandnew user, or adds admin permissions to an existing accounf. The
'delete-admin' script just deletes a user (maybe I need to rename it for
clarity). The 'list-admin' command list the current admins and displays
their user info.