Bound x509 cert credentials to django users!
This is a deceptively simple checkin, as it took me a solid dy to figure
out how to get the stars to align with the remote user authentication.

So first off, it is important to note that Django is not the smae as
DRF. DRF is an extension on top of django and adds other authentication
backends. See this for more info:

http://www.django-rest-framework.org/api-guide/authentication/#setting-the-authentication-scheme

So in previously looking at other django client auth modules, they
focused on providing capability to django itself, and they weren't
compatible with DRF. The simplest way to perform x509 client auth without
writing a DRF plugin (note - thats a good idea for another project!) was
to use RemoteUserAuthentication. Ironically, DRF doesnt need any
configuration to automagically rely on Django's authetication middleware.

Django relies on the REMOTE_USER header being set, which can be done by
Nginx passing via wsgi. However since I'm passing requests to DRF via
HTTP I had to create a custom middle ware class (that I'll eventually
extend) to use the prexiously set HTTP_X_SSL_CLIENT_S_DN header.

Once all of the configuration is in place, the user access the item
service page, and they're automatically logged in. Constantly. You can't
really log out. You are automatically logged in / authorized based on
the passed in credentials. In effect this makes it so that anyone with a
client cert can automatically log in, which makes for a very user
friendly experience (although potentially unsecure).

- Created the 'x509_middleware.py' file with the custom authntication
  middleware class 'ReverseProxyAuthMiddleware' which ovverides the
  header that is used to determine the username.
- Added the appropriate middleware and authentication backend settings
  to setting.py, which allows for automatic user association and login.
- Removed the 'rest_framework.urls' from the url list. This native URL
  set was used for logging in / out, but is no longer required (and in
  fact isn't really functional) with the changes to use remote auth.