Validated commands and switched to https for EC2
I validated the commands in the documentation by re-setting up a
snowball edge. As it turns out a lot of stuff changes (certs, arns,
various ids) when you reboot it. There isn't a lot of static
configuration on the snowball edge, and you will basically need to
re-configure it on every reboot.

I got the aws cli profile correctly configured with the snowball
'ca_bundle' so that all of the aws ec2 commands against the snowball
should be using https.