Added an inventory class tied to Items and UICs
This new Inventory model class has foreign key relations to the Item and
UnitIdentificationCode model classes. Inventory is assigned per UIC and
tracked seperatley. I added the same standard views and serializer.

The has a couple of tweaks though. Django unfortunatley doesn't support
composite keys, so... I basically updated the serializer to enforce that
created keys are a composite of the associated UIC and NSN. This is a
less than ideal way to do it, but it will work for the time constraints
I have for my demo. That should ensure that primary keys don't collide
for replication purposes.

Ideally you'd probably have a default id set in the model class that
calls a generator function. And based on some kind of configuration it
would create a unique record for each site (possibly not tied to a UIC).
Moving away from using natural keys entirely would ensure that collisions
are much more infrequent.

I also left in soem commented out code that I used to overwrite the
serializer create function. But withoput overriding the validate
function it was throwing errors, that should have never ocurred in the
first place.