Added fine grain permission controls
Added the rest_framework.permissions.DjangoModelPermissions class so
that DRF can manage user access to views based on crud roles. The roles
are generated by the generic views, and it makes user authorization easy
peasy.

Updated all of the views to require permissions.DjangoModelPermissions,
so that users that aren't in a specific group can't do anything to data.
Made sure that the User data view is locked down to admins only.

This checkin actually causes another issue that I'll fix shortly. The
way the certs are generated, the DRF user management panel complains
about having spaces in the username. I'll need to alter the remote auth
backend to clean the usernames up and simplify things.