Cleaned Remote Auth provided usernames
Cert subjects have a lot of characters in them that django just doesnt
like to have in usernames. Sure you can use the model classes and add
them in directly, but thats a pain. The UI complains about almost all
special chars.

Obviously this change wouldn't really fly in a real production
enviroment, but again, it's another simplification for the demo.

I created a RemoteUserBackend class  and overrode the clean_username
function. Basically it pull the user's actual name off the cert subject
and strips all special chars. This is ripe for username coliisions.

Updated the django settinsg to use the new ReverseProxyRemoteUserBackend