Added item data, updated item model
Added item reference data in CSV format, and the commands to load it
into the database.

The item model required a chnage to the precision of the price, loading
the bata balked on a price (10534.24) that was over the 6 digit
precision. I bumped it up to an arbitrary number, no real need to
optimize. Added the migration.