Reduced item reference data to increase page speed
I reduced the amount of item data (from 13000 rows to 1300 rows) in
order to speed up the inventory page load times. That page is slow to
load because it pulls all of the item and uic data and feeds it into a
dropdown box.

I separated the item and unitofissue reference data, and deleted the old
giant massive ref data blob.

Updated the query notes to include commands to delet the last 1000 rows
from the item table (used to prune the data) and the command to truncate
the item data (useful for reload)